requirements.txt
Requests==2.32.3
python-dotenv~=0.21.0
progress~=1.6
gitpython~=3.1.43